+++ HISTORY.md
@@ -1,6 +1,15 @@
 # Release History
 
 
+## 0.2.0 / 2012-02-01
+
+This release adds support for $stdout and $stderr capturing.
+
+Changes:
+
+* Handle captured stdout and stderr.
+
+
 ## 0.1.0 / 2011-12-13
 
 This is the first release of `RSpec On Tap`, a formatter for RSpec